Move graph serialization templates out of tt_metal public API
Move serialize_member and serialize_tracked_arg from graph_tracking.hpp
into ttnn/api/ttnn/graph/graph_serialization.hpp. These templates
depend on <reflect> and tt_stl/reflection.hpp, which should not be in
tt_metal's public API.

Changes:
- graph_tracking.hpp: replace nlohmann/json.hpp with json_fwd.hpp,
  keep only a forward declaration of serialize_tracked_arg, move
  IGraphProcessor::end_capture() default body to graph_tracking.cpp
- Create ttnn/api/ttnn/graph/graph_serialization.hpp with the full
  template definitions
- Update callers (tensor_ops.cpp, to_string.cpp, pytensor.cpp,
  py_to_tt_tensor.cpp, device_operation.hpp) to include the new header
