+++ tt_metal/api/tt-metalium/graph_tracking.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include <nlohmann/json.hpp>
+#include <nlohmann/json_fwd.hpp>
 #include <stdint.h>
 #include <any>
 #include <array>
@@ -12,7 +12,6 @@
 #include <memory>
 #include <mutex>
 #include <span>
-#include <sstream>
 #include <string>
 #include <string_view>
 #include <unordered_set>
@@ -36,92 +35,11 @@ struct TrackedArgument {
     std::string (*to_string_fn)(const std::any&);
 };
 
-// Serialization helpers for graph argument tracking.
-// The templates below rely on symbols from tt_stl/reflection.hpp (ttsl::reflection,
-// ttsl::is_specialization_v, ttsl::concepts::Reflectable, reflect::*).  These are
-// intentionally NOT included here to avoid pulling heavyweight headers into every
-// translation unit that includes graph_tracking.hpp.  The symbols are resolved at
-// template instantiation time via transitive includes in the calling TUs.
-namespace graph_detail {
-
-template <typename MemberT>
-void serialize_member(std::ostringstream& oss, const MemberT& member) {
-    if constexpr (std::is_enum_v<MemberT>) {
-        ttsl::reflection::operator<<(oss, member);
-    } else if constexpr (ttsl::is_specialization_v<MemberT, std::vector>) {
-        oss << "{";
-        for (size_t i = 0; i < member.size(); ++i) {
-            if (i > 0) {
-                oss << ", ";
-            }
-            serialize_member(oss, member[i]);
-        }
-        oss << "}";
-    } else if constexpr (ttsl::is_specialization_v<MemberT, std::optional>) {
-        if (member.has_value()) {
-            serialize_member(oss, member.value());
-        } else {
-            oss << "std::nullopt";
-        }
-    } else if constexpr (ttsl::is_specialization_v<MemberT, std::reference_wrapper>) {
-        serialize_member(oss, member.get());
-    } else if constexpr (ttsl::is_specialization_v<MemberT, std::pair>) {
-        oss << "{";
-        serialize_member(oss, member.first);
-        oss << ", ";
-        serialize_member(oss, member.second);
-        oss << "}";
-    } else if constexpr (requires { oss << member; }) {
-        oss << member;
-    } else if constexpr (ttsl::concepts::Reflectable<MemberT>) {
-        oss << reflect::type_name<MemberT>() << "(";
-        reflect::for_each(
-            [&oss, &member](auto I) {
-                if constexpr (I > 0) {
-                    oss << ", ";
-                }
-                serialize_member(oss, reflect::get<I>(member));
-            },
-            member);
-        oss << ")";
-    } else {
-        oss << "<" << reflect::type_name<MemberT>() << ">";
-    }
-}
-
-}  // namespace graph_detail
-
+// Forward declaration only – the definition lives in ttnn/graph/graph_serialization.hpp
+// which pulls in <reflect> and tt_stl/reflection.hpp.  This keeps the public API header
+// free of those heavyweight dependencies.
 template <typename T>
-std::string serialize_tracked_arg(const std::any& a) {
-    std::ostringstream oss;
-    const auto& ref = std::any_cast<const std::reference_wrapper<T>&>(a);
-    const auto& val = ref.get();
-    using CleanT = std::remove_cv_t<T>;
-
-    // Guard against incomplete types (e.g. forward-declared MeshCommandQueue)
-    // so that type traits below are never evaluated on them.
-    if constexpr (!requires { sizeof(CleanT); }) {
-        oss << "<incomplete type>";
-    } else if constexpr (ttsl::is_specialization_v<CleanT, std::vector>) {
-        ttsl::reflection::operator<<(oss, val);
-    } else if constexpr (requires { oss << val; }) {
-        oss << val;
-    } else if constexpr (ttsl::concepts::Reflectable<CleanT>) {
-        oss << reflect::type_name<CleanT>() << "(";
-        reflect::for_each(
-            [&oss, &val](auto I) {
-                if constexpr (I > 0) {
-                    oss << ", ";
-                }
-                graph_detail::serialize_member(oss, reflect::get<I>(val));
-            },
-            val);
-        oss << ")";
-    } else {
-        oss << "<" << reflect::type_name<T>() << ">";
-    }
-    return oss.str();
-}
+std::string serialize_tracked_arg(const std::any& a);
 
 class IGraphProcessor {
 public:
@@ -155,7 +73,7 @@ public:
 
     virtual void begin_capture(RunMode /*mode*/){};
 
-    virtual nlohmann::json end_capture() { return nullptr; };
+    virtual nlohmann::json end_capture();
 
     virtual ~IGraphProcessor() = default;
 };
